Merge export redesign branch back into trunk.
Includes a completely redesign approach to all export data types based on the use of NSOperation subclasses. CSV, SQL, XML and dot export types are currently functional, while the source files for PDF and HTML export types exist they are to be implemented, but are currently hidden from the interface. 

Also includes the following:

- Completely redesigned export interface.
    - The ability to customize CSV NULL values.
    - The ability to specify whether the UTF-8 BOM should be used in SQL dumps.
    - The ability to specify whether BLOB fields are output as hex or plain text during SQL dumps. Defaults to hex.

- Exporting currently selected tables via the tables list context menu.

Outstanding issues:

- Not all progress indicators for all export types are functional (or functioning correctly).
- A few issues related to the introduction of only exporting the content and create and drop syntax of specific tables during SQL dumps.

Needs some serious testing and benchmarking to ensure it replicates the current export functionality.