• initial preparations to support the storage of the entire SP session (all windows including tabs) • added file extension 'spfs' as SP bundle
Note: The idea is to save inside the given spfs bundle a file 'Info.list' which contains the entire structure (which window, which tabs, selected items, etc.; each single connection is saved as separate spf file - if untitled inside the bundle in the sub-folder 'Contents' or if not the absolute path to a already saved spf file). This should insure that the user can open a single spf file in SP or as part of a spfs bundle session without loosing having two different spf file for the same connection. - Comments are welcome